requirements.txt
uvicorn
gunicorn
openai
requests
python-dotenv
fastapi[standard]